Add files via upload
CCD dump from a 1994 Chrysler Voyager european version made with a Hantek 6022BL.
There are separate dumps for the mechanical instrument cluster and the digital instrument cluster.
The body controller from the digital instrument cluster stores the mileage, so there is a possibility the mileage is transmitted via the CCD bus. This should not be the case with the mechanical instrument cluster.

The dumps are created with PulseView with with 2M samples at a rate of 1mhz. Other values would not work reliably.